Refactor k-bar
