+++ panel/src/config/components.js
@@ -200,7 +200,7 @@ Vue.component("k-url-field-preview", UrlFieldPreview);
 Vue.component("k-users-field-preview", UsersFieldPreview);
 
 /* Layout */
-import Bar from "@/components/Layout/Bar.vue";
+import Bar from "@/ui/components/Bar.vue";
 import Box from "@/components/Layout/Box.vue";
 import Card from "@/components/Layout/Card.vue";
 import Cards from "@/components/Layout/Cards.vue";
