+++ panel/src/ui/components/Bar.vue
@@ -1,8 +1,26 @@
 <template>
   <div class="k-bar">
-    <div v-if="$slots.left" class="k-bar-slot" data-position="left"><slot name="left" /></div>
-    <div v-if="$slots.center" class="k-bar-slot" data-position="center"><slot name="center" /></div>
-    <div v-if="$slots.right" class="k-bar-slot" data-position="right"><slot name="right" /></div>
+    <div
+      v-if="$slots.left"
+      class="k-bar-slot"
+      data-position="left"
+    >
+      <slot name="left" />
+    </div>
+    <div
+      v-if="$slots.center"
+      class="k-bar-slot"
+      data-position="center"
+    >
+      <slot name="center" />
+    </div>
+    <div
+      v-if="$slots.right"
+      class="k-bar-slot"
+      data-position="right"
+    >
+      <slot name="right" />
+    </div>
   </div>
 </template>
